requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-email-confirmation==0.2
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-authtoken==1.2.4
django-uni-form==0.9.0
djangorestframework==3.11.0
filelock==3.0.12
github.py==0.5.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
multidict==4.7.6
oauth2==1.9.0.post1
oauthlib==3.1.0
pbr==5.4.5
Pillow==8.0.1
psycopg2==2.8.5
pycodestyle==2.6.0
python-decouple==3.3
python-openid==2.2.5
python3-openid==3.2.0
pytz==2020.1
qrcode==6.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
stevedore==3.2.0
toml==0.10.2
Unipath==1.1
urllib3==1.25.9
virtualenv==20.0.27
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
whitenoise==5.2.0
yarl==1.4.2
